tv-chart: apply clipped candle scale to all CandleSeries
