+++ src/lib/charts/candle-data-feed.svelte.ts
@@ -1,6 +1,5 @@
-import type { AutoscaleInfo, UTCTimestamp } from 'lightweight-charts';
-import type { ApiCandle, CandleDataItem, CandleTimeBucket, DataFeed, TvDataItem } from './types';
-import { chartWickThreshold } from '$lib/config';
+import type { UTCTimestamp } from 'lightweight-charts';
+import type { ApiCandle, CandleDataItem, CandleTimeBucket, DataFeed } from './types';
 import { type TimeInterval, utcMinute, utcHour, utcDay } from 'd3-time';
 import { fetchPublicApi } from '$lib/helpers/public-api';
 
@@ -16,8 +15,6 @@ export function tsToUnixTimestamp(ts: string) {
 	return (new Date(`${ts}Z`).valueOf() / 1000) as UTCTimestamp;
 }
 
-export type TransformItem = (c: ApiCandle) => CandleDataItem;
-
 export function apiCandleToDataItem(c: ApiCandle): CandleDataItem {
 	return {
 		time: tsToUnixTimestamp(c.ts),
@@ -29,36 +26,6 @@ export function apiCandleToDataItem(c: ApiCandle): CandleDataItem {
 }
 
 export type ApiDataTransformer = (data: any) => CandleDataItem[];
-
-/**
- * Custom PriceScaleCalculator for trading pair price candle data
- *
- * By default, TradingView CandlestickSeries determines price scale using min/max
- * of all currently displayed candle values. For trading pairs with extremely long
- * wicks (high/low values), this results in an overly flattened-out price scale.
- * E.g.: http://localhost:5173/trading-view/arbitrum/uniswap-v3/crv-eth-fee-30
- *
- * calculateClippedCandleScale address this by clipping the high/low values based on a
- * configurable threshold.
- */
-export function calculateClippedCandleScale(candles: TvDataItem[]): AutoscaleInfo | null {
-	if (candles.length === 0) return null;
-
-	const priceRange = (candles as CandleDataItem[]).reduce(
-		({ minValue, maxValue }, { open, high, low, close }) => {
-			const clippedLow = Math.max(low, Math.min(open, close) * (1 - chartWickThreshold));
-			const clippedHigh = Math.min(high, Math.max(open, close) * (1 + chartWickThreshold));
-			return {
-				minValue: Math.min(minValue, clippedLow),
-				maxValue: Math.max(maxValue, clippedHigh)
-			};
-		},
-		// initial accumulator: ±Infinity ensures any candle value will be lower/higher
-		{ minValue: Infinity, maxValue: -Infinity }
-	);
-
-	return { priceRange };
-}
 
 export class CandleDataFeed implements DataFeed<CandleDataItem> {
 	static timeBuckets: CandleTimeBucket[] = ['1m', '5m', '15m', '1h', '4h', '1d', '7d', '30d'];
